src: Allow clippy::needless_range_loop
These functions allocate the return array ahead of time and use
macros to convert C values to Rust values, which makes addressing
the warnings reported by clippy non trivial.

The underlying libvirt APIs are considered deprecated, so in the
long run it's probably better to rework the code more extensively
or even drop it entirely.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>