ci: Drop old targets
libvirt has already dropped support for Ubuntu 18.04, so no
problem there.

We are however dropping support for Debian 10 ahead of the main
library, because the Rust version there is too old to implement
some features that are needed by dependencies that we're about
to introduce.

Debian 10 goes EOL in August 2022, so we're only cutting support
short by a few months.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>