util: Add macros to generate enums
These macros generate the boilerplate for conversion between libvirt C
enums and Rust enums. The inner macros are implemented as push-down
accumulators so that we can incrementally parse the match arms and
generate a match expression in one shot. This step is necessary because
Rust macros can only output valid Rust syntax. See
https://danielkeep.github.io/tlborm/book/pat-push-down-accumulation.html
for details.

Signed-off-by: Wim de With <wf@dewith.io>