connect: Provide an example to open() in its docs string
Just above the Connect::open() implementation there's its
documentation. But since this is the most important function
there is, provide an example on how to use it. We already have
more examples in the top most documentation (src/lib.rs) but it
doesn't hurt to have one here too.

This also silences one of 'cargo doc' warnings.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>