src: mark all constructors taking C pointers as unsafe
Almost all the 'new' and 'from_ptr' methods are taking raw C
pointers to underlying libvirt C objects. This is an unsafe
design as it relies on the caller passing a valid C pointer.

While a few of the 'from_ptr' methods were already decorated
with 'unsafe', the 'new' methods were not.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>