src: rename 'new' methods to 'from_ptr'
When 'new' methods are all taking a raw C pointer to the underlying
libvirt object. They are the natural counterpart to 'as_ptr' and as
such the normal rust naming convention is 'from_ptr'.

This is an API break, however, few (if any) apps using libvirt
would ever need to call the 'new' methods, so the impact should be
negligible. They would have to be interacting with another C library
that deals with raw libvirt C object pointers to need to use 'new'.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>