domain: fix passing of typed parameters to C API
When we have the pattern

   let something: Vec<Thing> = Vec::with_capacity(size);

   SomeAPI(&something[0]);

The use of &something[0] will trigger a panic at runtime. The
'with_capacity' method only pre-allocates memory for the vector,
it does not set the length of the vector, which remains at zero.
Thus attempting to access the first elemnt via [0] is invalid.

The correct thing todo is to access the underlying memory of
the vector, instead of the address of the first (non-existant)
element.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>