src: fix ref counting in stream event callback
When a stream event callback is invoked, it is passed a virStreamPtr
instance as one of its parameters.

The Rust callback impl creates a Rust "Stream" object around this
pointer. The callback does not own any reference on this pointer
though. So when the Rust "Stream" object is dropped, it will release
a reference on the virStreamPtr that it never owned. This eventually
leads to crashes when an already free'd virStreamPtr is used.

https://gitlab.com/libvirt/libvirt-rust/-/issues/2
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>