src: fix UAF of the stream pointer in drop method
The drop method for Stream attempts to drop its pointer first and then
its callback. However, event_remove_callback(), which is called to drop
the callback, needs to reference to the pointer. This results in UAF of
the stream pointer and causes a panic, So the callback should be dropped
first.

Signed-off-by: Ryosuke Yasuoka <ryasuoka@redhat.com>