src: mark all 'as_ptr' methods as unsafe
The 'as_ptr' method gives the caller access to the underlying C handle
backing the object. If the caller were to release the reference on this
handle, then the Rust wrapper is liable to end up holding a pointer to
a released object. This is an unsafe scenario, so the 'as_ptr' method
must be declared unsafe.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>