src: implement the Drop trait for Connect objects
The C virConnectClose method (and thus Rust close() method) are
misleadingly named. While they're said to close the connection,
what they actually do is decrement a reference counter. When the
last reference is dropped, the connection is then closed.

The Rust close() method will set self.ptr to None, which means
any further methods on the Connect object will panic referencing
self.ptr

Releasing a reference is something that should be done implicitly
when an object goes out of scope, which is the purpose of the Drop
trait. Implement Drop and mark the close method as private and
unsafe. Typical usage of the Connect object can now avoid mutable
variables.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>