src: mark 'free' methods private and unsafe
The 'free' method will directly release a reference on the underlying
C object. Once this is done any other method calls will panic trying
to reference self.ptr which is now None. IOW, the 'free' method should
be marked unsafe. There is also no reason to call 'free' directly since
the Drop trait is implemented.

Using the 'drop' method instead of 'free' also means that the varibles
do not have to be declared mutable, which triggers some cleanup in the
tests.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>