Add missing refs on object back pointers
The 'virXXXXGetConnect' methods, and similar APIs that return back
pointers to the parent objects all return a borrowed reference. If
the caller wants to retain the pointer they must acquire their own
explicit reference. The rust code failed todo this, which would
result in a reference count underflow when the Connect is dropped,
and likely use-after-fere.

Reported-by: Wim de With <wf@dewith.io>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>